# [Merged by Bors] - Faster view frustum culling