--- a/crates/bevy_pbr/src/light.rs
+++ b/crates/bevy_pbr/src/light.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 
 use bevy_asset::Assets;
 use bevy_ecs::prelude::*;
-use bevy_math::{Mat4, UVec2, UVec3, Vec2, Vec3, Vec3Swizzles, Vec4, Vec4Swizzles};
+use bevy_math::{Mat4, UVec2, UVec3, Vec2, Vec3, Vec3A, Vec3Swizzles, Vec4, Vec4Swizzles};
 use bevy_reflect::Reflect;
 use bevy_render::{
     camera::{Camera, CameraProjection, OrthographicProjection},
@@ -640,8 +640,8 @@ fn cluster_space_light_aabb(
     light_sphere: &Sphere,
 ) -> (Vec3, Vec3) {
     let light_aabb_view = Aabb {
-        center: (inverse_view_transform * light_sphere.center.extend(1.0)).xyz(),
-        half_extents: Vec3::splat(light_sphere.radius),
+        center: Vec3A::from(inverse_view_transform * light_sphere.center.extend(1.0)),
+        half_extents: Vec3A::splat(light_sphere.radius),
     };
     let (mut light_aabb_view_min, mut light_aabb_view_max) =
         (light_aabb_view.min(), light_aabb_view.max());
@@ -798,7 +798,7 @@ pub(crate) fn assign_lights_to_clusters(
                 false
             } else {
                 let light_sphere = Sphere {
-                    center: light.translation,
+                    center: Vec3A::from(light.translation),
                     radius: light.range,
                 };
 
@@ -875,7 +875,7 @@ pub(crate) fn assign_lights_to_clusters(
             let mut cluster_index_estimate = 0.0;
             for light in lights.iter() {
                 let light_sphere = Sphere {
-                    center: light.translation,
+                    center: Vec3A::from(light.translation),
                     radius: light.range,
                 };
 
@@ -965,7 +965,7 @@ pub(crate) fn assign_lights_to_clusters(
 
         for light in lights.iter() {
             let light_sphere = Sphere {
-                center: light.translation,
+                center: Vec3A::from(light.translation),
                 radius: light.range,
             };
 
@@ -1209,7 +1209,7 @@ pub fn check_light_mesh_visibility(
 
                 let view_mask = maybe_view_mask.copied().unwrap_or_default();
                 let light_sphere = Sphere {
-                    center: transform.translation,
+                    center: Vec3A::from(transform.translation),
                     radius: point_light.range,
                 };