--- a/crates/bevy_render/src/primitives/mod.rs
+++ b/crates/bevy_render/src/primitives/mod.rs
@@ -6,12 +6,14 @@ use bevy_reflect::Reflect;
 #[derive(Component, Clone, Debug, Default, Reflect)]
 #[reflect(Component)]
 pub struct Aabb {
-    pub center: Vec3,
-    pub half_extents: Vec3,
+    pub center: Vec3A,
+    pub half_extents: Vec3A,
 }
 
 impl Aabb {
     pub fn from_min_max(minimum: Vec3, maximum: Vec3) -> Self {
+        let minimum = Vec3A::from(minimum);
+        let maximum = Vec3A::from(maximum);
         let center = 0.5 * (maximum + minimum);
         let half_extents = 0.5 * (maximum - minimum);
         Self {
@@ -23,7 +25,7 @@ impl Aabb {
     /// Calculate the relative radius of the AABB with respect to a plane
     pub fn relative_radius(&self, p_normal: &Vec3A, axes: &[Vec3A]) -> f32 {
         // NOTE: dot products on Vec3A use SIMD and even with the overhead of conversion are net faster than Vec3
-        let half_extents = Vec3A::from(self.half_extents);
+        let half_extents = self.half_extents;
         Vec3A::new(
             p_normal.dot(axes[0]),
             p_normal.dot(axes[1]),
@@ -33,11 +35,11 @@ impl Aabb {
         .dot(half_extents)
     }
 
-    pub fn min(&self) -> Vec3 {
+    pub fn min(&self) -> Vec3A {
         self.center - self.half_extents
     }
 
-    pub fn max(&self) -> Vec3 {
+    pub fn max(&self) -> Vec3A {
         self.center + self.half_extents
     }
 }
@@ -46,14 +48,14 @@ impl From<Sphere> for Aabb {
     fn from(sphere: Sphere) -> Self {
         Self {
             center: sphere.center,
-            half_extents: Vec3::splat(sphere.radius),
+            half_extents: Vec3A::splat(sphere.radius),
         }
     }
 }
 
 #[derive(Debug, Default)]
 pub struct Sphere {
-    pub center: Vec3,
+    pub center: Vec3A,
     pub radius: f32,
 }
 
@@ -65,7 +67,7 @@ impl Sphere {
             Vec3A::from(local_to_world.y_axis),
             Vec3A::from(local_to_world.z_axis),
         ];
-        let v = Vec3A::from(aabb_center_world) - Vec3A::from(self.center);
+        let v = Vec3A::from(aabb_center_world) - self.center;
         let d = v.length();
         let relative_radius = aabb.relative_radius(&(v / d), &axes);
         d < self.radius + relative_radius
@@ -96,8 +98,8 @@ impl Plane {
 
     /// `Plane` unit normal
     #[inline]
-    pub fn normal(&self) -> Vec3 {
-        self.normal_d.xyz()
+    pub fn normal(&self) -> Vec3A {
+        Vec3A::from(self.normal_d)
     }
 
     /// Signed distance from the origin along the unit normal such that n.p + d = 0 for point p in
@@ -215,7 +217,7 @@ mod tests {
         // Sphere outside frustum
         let frustum = big_frustum();
         let sphere = Sphere {
-            center: Vec3::new(0.9167, 0.0000, 0.0000),
+            center: Vec3A::new(0.9167, 0.0000, 0.0000),
             radius: 0.7500,
         };
         assert!(!frustum.intersects_sphere(&sphere));
@@ -226,7 +228,7 @@ mod tests {
         // Sphere intersects frustum boundary
         let frustum = big_frustum();
         let sphere = Sphere {
-            center: Vec3::new(7.9288, 0.0000, 2.9728),
+            center: Vec3A::new(7.9288, 0.0000, 2.9728),
             radius: 2.0000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -251,7 +253,7 @@ mod tests {
         // Sphere surrounds frustum
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(0.0000, 0.0000, 0.0000),
+            center: Vec3A::new(0.0000, 0.0000, 0.0000),
             radius: 3.0000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -262,7 +264,7 @@ mod tests {
         // Sphere is contained in frustum
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(0.0000, 0.0000, 0.0000),
+            center: Vec3A::new(0.0000, 0.0000, 0.0000),
             radius: 0.7000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -273,7 +275,7 @@ mod tests {
         // Sphere intersects a plane
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(0.0000, 0.0000, 0.9695),
+            center: Vec3A::new(0.0000, 0.0000, 0.9695),
             radius: 0.7000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -284,7 +286,7 @@ mod tests {
         // Sphere intersects 2 planes
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(1.2037, 0.0000, 0.9695),
+            center: Vec3A::new(1.2037, 0.0000, 0.9695),
             radius: 0.7000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -295,7 +297,7 @@ mod tests {
         // Sphere intersects 3 planes
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(1.2037, -1.0988, 0.9695),
+            center: Vec3A::new(1.2037, -1.0988, 0.9695),
             radius: 0.7000,
         };
         assert!(frustum.intersects_sphere(&sphere));
@@ -306,7 +308,7 @@ mod tests {
         // Sphere avoids intersecting the frustum by 1 plane
         let frustum = frustum();
         let sphere = Sphere {
-            center: Vec3::new(-1.7020, 0.0000, 0.0000),
+            center: Vec3A::new(-1.7020, 0.0000, 0.0000),
             radius: 0.7000,
         };
         assert!(!frustum.intersects_sphere(&sphere));
@@ -331,7 +333,7 @@ mod tests {
         // Sphere outside frustum
         let frustum = long_frustum();
         let sphere = Sphere {
-            center: Vec3::new(-4.4889, 46.9021, 0.0000),
+            center: Vec3A::new(-4.4889, 46.9021, 0.0000),
             radius: 0.7500,
         };
         assert!(!frustum.intersects_sphere(&sphere));
@@ -342,7 +344,7 @@ mod tests {
         // Sphere intersects frustum boundary
         let frustum = long_frustum();
         let sphere = Sphere {
-            center: Vec3::new(-4.9957, 0.0000, -0.7396),
+            center: Vec3A::new(-4.9957, 0.0000, -0.7396),
             radius: 4.4094,
         };
         assert!(frustum.intersects_sphere(&sphere));