--- a/crates/bevy_render/src/primitives/mod.rs
+++ b/crates/bevy_render/src/primitives/mod.rs
@@ -159,8 +159,9 @@ impl Frustum {
 
     #[inline]
     pub fn intersects_sphere(&self, sphere: &Sphere) -> bool {
+        let sphere_center = sphere.center.extend(1.0);
         for plane in &self.planes {
-            if plane.normal_d().dot(sphere.center.extend(1.0)) + sphere.radius <= 0.0 {
+            if plane.normal_d().dot(sphere_center) + sphere.radius <= 0.0 {
                 return false;
             }
         }
@@ -169,7 +170,7 @@ impl Frustum {
 
     #[inline]
     pub fn intersects_obb(&self, aabb: &Aabb, model_to_world: &Mat4) -> bool {
-        let aabb_center_world = *model_to_world * aabb.center.extend(1.0);
+        let aabb_center_world = model_to_world.transform_point3a(aabb.center).extend(1.0);
         let axes = [
             Vec3A::from(model_to_world.x_axis),
             Vec3A::from(model_to_world.y_axis),